Refactor mysql cdc test case.
