+++ bitsail-test/bitsail-test-integration/bitsail-test-integration-cdc/src/test/java/com/bytedance/bitsail/test/integration/cdc/mysql/container/MySQLContainerMariadbAdapter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.bytedance.bitsail.test.integration.container;
+package com.bytedance.bitsail.test.integration.cdc.mysql.container;
 
 import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.utility.DockerImageName;
